+++ acl/tests/test_model.py
@@ -2,6 +2,7 @@ from django.test import TestCase
 from django.contrib.auth.models import Permission
 from group.models import Group
 from acl.models import ACLBase
+from role.models import Role
 from user.models import User
 from importlib import import_module
 from airone.lib.acl import ACLType
@@ -10,6 +11,10 @@ from airone.lib.acl import ACLType
 class ModelTest(TestCase):
     def setUp(self):
         self.user = User.objects.create(username="foo", email="hoge@example.com", password="fuga")
+
+        # create test Role instance and set user to be belonged to it
+        self.role = Role.objects.create(name="test_role")
+        self.role.users.add(self.user)
 
     def test_acl_base(self):
         # chacks to enable embedded acl field
@@ -42,7 +47,7 @@ class ModelTest(TestCase):
         aclobj = ACLBase.objects.create(name="hoge", created_user=self.user, is_public=False)
 
         # set correct permission
-        self.user.permissions.add(aclobj.readable)
+        self.role.permissions.add(aclobj.readable)
 
         self.assertTrue(self.user.has_permission(aclobj, ACLType.Readable))
 
@@ -50,7 +55,7 @@ class ModelTest(TestCase):
         aclobj = ACLBase.objects.create(name="hoge", created_user=self.user, is_public=False)
 
         # set surperior permission
-        self.user.permissions.add(aclobj.writable)
+        self.role.permissions.add(aclobj.writable)
 
         self.assertTrue(self.user.has_permission(aclobj, ACLType.Readable))
 
@@ -60,37 +65,22 @@ class ModelTest(TestCase):
         aclobj = ACLBase.objects.create(name="hoge", created_user=self.user, is_public=False)
 
         # set inferior permission
-        self.user.permissions.add(aclobj.readable)
+        self.role.permissions.add(aclobj.readable)
 
         self.assertFalse(another_user.has_permission(aclobj, ACLType.Writable))
 
     def test_pass_permission_check_with_group_permissoin(self):
         another_user = User.objects.create(username="bar", email="bar@example.com", password="")
         group = Group.objects.create(name="hoge")
+        self.role.groups.add(group)
 
         aclobj = ACLBase.objects.create(name="hoge", created_user=self.user, is_public=False)
 
         # set correct permission to the group that test user is belonged to
-        group.permissions.add(aclobj.readable)
+        self.role.permissions.add(aclobj.readable)
         another_user.groups.add(group)
 
         self.assertTrue(another_user.has_permission(aclobj, ACLType.Readable))
-
-    def test_get_registered_user_permissoins(self):
-        aclobj = ACLBase.objects.create(name="hoge", created_user=self.user, is_public=False)
-        self.user.permissions.add(aclobj.readable)
-
-        self.assertEqual(self.user.get_acls(aclobj).count(), 1)
-        self.assertEqual(self.user.get_acls(aclobj)[0], aclobj.readable)
-
-    def test_get_registered_group_permissoins(self):
-        group = Group.objects.create(name="hoge")
-
-        aclobj = ACLBase.objects.create(name="hoge", created_user=self.user, is_public=False)
-        group.permissions.add(aclobj.full)
-
-        self.assertEqual(group.get_acls(aclobj).count(), 1)
-        self.assertEqual(group.get_acls(aclobj)[0], aclobj.full)
 
     def test_get_subclass_object(self):
         # make objects to test
@@ -207,155 +197,3 @@ class ModelTest(TestCase):
 
         self.assertTrue(superuser.has_permission(aclobj, ACLType.Full))
         self.assertFalse(guestuser.has_permission(aclobj, ACLType.Full))
-
-    def test_check_may_permitted(self):
-        admin_user = User.objects.create(username="admin", is_superuser=True)
-        non_admin_user = User.objects.create(username="user", is_superuser=False)
-        aclobj = ACLBase.objects.create(name="hoge", created_user=self.user, is_public=False)
-
-        # add and register group to non_admin_user
-        group = Group.objects.create(name="group", is_active=True)
-        non_admin_user.groups.add(group)
-
-        acl_bases = {
-            "acl1": {"is_public": True, "default_permission": ACLType.Nothing.id},
-            "acl2": {"is_public": False, "default_permission": ACLType.Full.id},
-            "acl3": {"is_public": False, "default_permission": ACLType.Readable.id},
-            "acl4": {"is_public": False, "default_permission": ACLType.Nothing.id},
-        }
-
-        acl = {
-            "user_readable": {
-                "acl_settings": [
-                    {
-                        "member_id": str(non_admin_user.id),
-                        "member_type": "user",
-                        "value": str(ACLType.Readable.id),
-                    }
-                ]
-            },
-            "user_full": {
-                "acl_settings": [
-                    {
-                        "member_id": str(non_admin_user.id),
-                        "member_type": "user",
-                        "value": str(ACLType.Full.id),
-                    }
-                ]
-            },
-            "group_full": {
-                "acl_settings": [
-                    {
-                        "member_id": str(group.id),
-                        "member_type": "group",
-                        "value": str(ACLType.Full.id),
-                    }
-                ]
-            },
-            "group_readable": {
-                "acl_settings": [
-                    {
-                        "member_id": str(group.id),
-                        "member_type": "group",
-                        "value": str(ACLType.Readable.id),
-                    }
-                ]
-            },
-            "nothing": {"acl_settings": []},
-        }
-
-        # checks that admin user can access any case
-        for info in acl_bases.values():
-            self.assertTrue(admin_user.may_permitted(aclobj, ACLType.Full, acl_settings=[], **info))
-
-        # checks permitted cases
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Readable, acl_settings=[], **acl_bases["acl1"]
-            )
-        )
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Readable, acl_settings=[], **acl_bases["acl2"]
-            )
-        )
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Readable, acl_settings=[], **acl_bases["acl3"]
-            )
-        )
-
-        param = {}
-        param["is_public"] = acl_bases["acl4"]["is_public"]
-        param["default_permission"] = acl_bases["acl4"]["default_permission"]
-
-        # checks permitted cases with indivisual permissions
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Readable, **dict(acl["user_readable"], **acl_bases["acl4"])
-            )
-        )
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["user_full"], **acl_bases["acl4"])
-            )
-        )
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["group_full"], **acl_bases["acl4"])
-            )
-        )
-
-        # checks unpermitted cases
-        self.assertFalse(
-            non_admin_user.may_permitted(aclobj, ACLType.Full, acl_settings=[], **acl_bases["acl3"])
-        )
-        self.assertFalse(
-            non_admin_user.may_permitted(aclobj, ACLType.Full, acl_settings=[], **acl_bases["acl4"])
-        )
-
-        # checks unpermitted cases with indivisual permissions
-        self.assertFalse(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["user_readable"], **acl_bases["acl4"])
-            )
-        )
-
-        # checks permitted cases with empty acl_settings parameters when user has permission
-        self.assertFalse(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["nothing"], **acl_bases["acl4"])
-            )
-        )
-        non_admin_user.permissions.add(aclobj.full)
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["nothing"], **acl_bases["acl4"])
-            )
-        )
-
-        # While user has full permission of object,
-        # this returns false when unpermitted input was specified
-        self.assertFalse(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["user_readable"], **acl_bases["acl4"])
-            )
-        )
-
-        # checks permitted cases with empty acl_settings parameters when group has permission
-        non_admin_user.permissions.remove(aclobj.full)
-        group.permissions.add(aclobj.full)
-        self.assertTrue(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["nothing"], **acl_bases["acl4"])
-            )
-        )
-
-        # While user has full permission of object,
-        # this returns false when unpermitted input was specified
-        self.assertFalse(
-            non_admin_user.may_permitted(
-                aclobj, ACLType.Full, **dict(acl["group_readable"], **acl_bases["acl4"])
-            )
-        )
-        non_admin_user.permissions.remove(aclobj.full)