+++ user/tests/test_model.py
@@ -5,6 +5,7 @@ from airone.lib.acl import ACLType
 from entity.models import Entity, EntityAttr
 from entry.models import Entry
 from group.models import Group
+from role.models import Role
 from user.models import User, History
 
 
@@ -107,8 +108,10 @@ class ModelTest(TestCase):
 
         user = User.objects.create(username="user")
         group = Group.objects.create(name="group")
+        role = Role.objects.create(name="role")
 
         user.groups.add(group)
+        role.groups.add(group)
 
         entity = Entity.objects.create(
             name="entity",
@@ -117,8 +120,24 @@ class ModelTest(TestCase):
             default_permission=ACLType.Nothing.id,
         )
 
-        group.permissions.add(entity.readable)
+        role.permissions.add(entity.readable)
 
         self.assertTrue(user.has_permission(entity, ACLType.Readable))
         self.assertFalse(user.has_permission(entity, ACLType.Writable))
         self.assertFalse(user.has_permission(entity, ACLType.Full))
+
+    def test_object_acl_that_should_not_be_shown(self):
+        user = User.objects.create(username="user")
+        entity = Entity.objects.create(
+            name="entity", created_user=user, is_public=False, default_permission=ACLType.Nothing.id
+        )
+
+        entry = Entry.objects.create(
+            name="Entry",
+            schema=entity,
+            is_public=False,
+            default_permission=ACLType.Full.id,
+            created_user=user,
+        )
+
+        self.assertFalse(user.has_permission(entry, ACLType.Readable))